Fix clang -Wc++11-narrowing warnings.
Clang compiler gives warning to implicit conversions of
literal structure fields, although the compiler generated
implicit conversions are what we want in these cases.

+++ camera/QCameraHWI_Parm.cpp
@@ -712,7 +712,7 @@ void QCameraHardwareInterface::loadTables()
 
 rat_t getRational(int num, int denom)
 {
-    rat_t temp = {num, denom};
+    rat_t temp = {(uint32_t)num, (uint32_t)denom};
     return temp;
 }
